top5/RCS_CNN_LSTM_Notebook/RCS_CNN_LSTM_Notebook/requirements.txt
MetaTrader5
matplotlib
numpy>=1.21.0,<3.0.0
onnxruntime-gpu>=1.12.0
onnx>=1.12.0
tf2onnx>=1.13.0
pandas>=1.3.0
scikit-learn>=1.0.0
shap>=0.40.0
streamlit>=1.15.0
tsfresh>=0.19.0
yfinance>=0.1.87
tensorflow-gpu>=2.8.0,<3.0.0
ta>=0.7.0
